move redis functionality to redis.go
